Move growthbook setup to growthbook provider
