+++ projects/bp-gallery/src/components/App.tsx
@@ -8,18 +8,13 @@ import AlertProvider from './provider/AlertProvider';
 import AuthProvider from './provider/AuthProvider';
 import ClipboardEditorProvider from './provider/ClipboardEditorProvider';
 import DialogProvider from './provider/DialogProvider';
+import { GrowthBookProvider } from './provider/GrowthBookProvider';
 import { MobileProvider } from './provider/MobileProvider';
 import { ScrollProvider } from './provider/ScrollProvider';
 import StorageProvider from './provider/StorageProvider';
 import routes from './routes';
 import BottomBar from './top-and-bottom-bar/BottomBar';
 import TopBar from './top-and-bottom-bar/TopBar';
-import { useEffect } from 'react';
-import { GrowthBook } from '@growthbook/growthbook-react';
-import { GrowthBookProvider } from './provider/GrowthBookProvider';
-
-const growthbookApiHost = import.meta.env.VITE_REACT_APP_GROWTHBOOK_APIHOST;
-const growthbookClientKey = import.meta.env.VITE_REACT_APP_GROWTHBOOK_CLIENTKEY;
 
 const apolloClient = new ApolloClient({
   link: buildHttpLink(sessionStorage.getItem('jwt')),
@@ -70,27 +65,7 @@ const apolloClient = new ApolloClient({
   }),
 });
 
-const growthbook = new GrowthBook({
-  apiHost: growthbookApiHost,
-  clientKey: growthbookClientKey,
-  enableDevMode: true,
-  trackingCallback: (experiment, result) => {
-    const w: any = window;
-    const _paq: Array<any> = (w._paq = w._paq || []);
-    _paq.push(['trackEvent', 'ExperimentViewed', experiment.key, 'v' + String(result.variationId)]);
-  },
-  onFeatureUsage: (featureKey, result) => {
-    const w: any = window;
-    const _paq: Array<any> = (w._paq = w._paq || []);
-    _paq.push(['trackEvent', 'ExperimentViewed', featureKey, 'v' + String(result)]);
-  },
-});
-
 const App = () => {
-  useEffect(() => {
-    growthbook.loadFeatures({ autoRefresh: true });
-  }, []);
-
   return (
     <ApolloProvider client={apolloClient}>
       <AlertProvider>
@@ -100,7 +75,7 @@ const App = () => {
               <MobileProvider>
                 <div className='App'>
                   <ClipboardEditorProvider>
-                    <GrowthBookProvider growthbook={growthbook}>
+                    <GrowthBookProvider>
                       <ScrollProvider useWindow>
                         <TopBar />
                         <ScrollContainer>{renderRoutes(routes)}</ScrollContainer>
