+++ projects/bp-gallery/src/components/provider/GrowthBookProvider.tsx
@@ -3,12 +3,30 @@ import {
   GrowthBookProvider as _GrowthBookProvider,
 } from '@growthbook/growthbook-react';
 import { PropsWithChildren, useEffect } from 'react';
-import { AppFeatures } from '../../helpers/growthbook';
 
-export const GrowthBookProvider = ({
-  children,
-  growthbook,
-}: PropsWithChildren<{ growthbook: GrowthBook<AppFeatures> | undefined }>) => {
+const growthbookApiHost = import.meta.env.VITE_REACT_APP_GROWTHBOOK_APIHOST;
+const growthbookClientKey = import.meta.env.VITE_REACT_APP_GROWTHBOOK_CLIENTKEY;
+
+const growthbook = new GrowthBook({
+  apiHost: growthbookApiHost,
+  clientKey: growthbookClientKey,
+  enableDevMode: true,
+  trackingCallback: (experiment, result) => {
+    const w: any = window;
+    const _paq: Array<any> = (w._paq = w._paq || []);
+    _paq.push(['trackEvent', 'ExperimentViewed', experiment.key, 'v' + String(result.variationId)]);
+  },
+  onFeatureUsage: (featureKey, result) => {
+    const w: any = window;
+    const _paq: Array<any> = (w._paq = w._paq || []);
+    _paq.push(['trackEvent', 'ExperimentViewed', featureKey, 'v' + String(result)]);
+  },
+});
+
+export const GrowthBookProvider = ({ children }: PropsWithChildren<{}>) => {
+  useEffect(() => {
+    growthbook.loadFeatures({ autoRefresh: true });
+  }, []);
   useEffect(() => {
     // Set user attributes for targeting (from cookie, auth system, etc.)
     let visitor_id: string;
@@ -18,12 +36,12 @@ export const GrowthBookProvider = ({
       function (this: any) {
         // eslint-disable-next-line @typescript-eslint/no-unsafe-call
         visitor_id = this.getVisitorId();
-        growthbook?.setAttributes({ ...growthbook.getAttributes(), id: visitor_id });
+        growthbook.setAttributes({ ...growthbook.getAttributes(), id: visitor_id });
       },
     ]);
 
     const refresh = (): NodeJS.Timeout => {
-      growthbook?.refreshFeatures();
+      growthbook.refreshFeatures();
       return setTimeout(refresh, 5000);
     };
